+++ frontend/src/components/matching/WaitingTimeRenderer.js
@@ -1,0 +1,31 @@
+import { useRef, useState } from "react";
+
+export default function WaitingTimeRenderer({ remainingTime }) {
+    const currentTime = useRef(remainingTime);
+    const isNewTimeFirstTick = useRef(false);
+    const [, setOneLastRerender] = useState(0);
+
+    if (currentTime.current !== remainingTime) {
+        isNewTimeFirstTick.current = true;
+        currentTime.current = remainingTime;
+    } else {
+        isNewTimeFirstTick.current = false;
+    }
+
+    // force one last re-render when the time is over to tirgger the last animation
+    if (remainingTime === 0) {
+        setTimeout(() => {
+            setOneLastRerender((val) => val + 1);
+        }, 20);
+    }
+
+    const isTimeUp = isNewTimeFirstTick.current;
+
+    return (
+        <div className="time-wrapper">
+            <div key={remainingTime} className={`time ${isTimeUp ? "up" : ""}`}>
+                {remainingTime}
+            </div>
+        </div>
+    );
+}
