+++ frontend/src/constants.js
@@ -1,3 +1,18 @@
-export const STATUS_CODE_CREATED = 201
-export const STATUS_CODE_SUCCESS = 200
-export const STATUS_CODE_CONFLICT = 409
+export const STATUS_CODE_CREATED = 201;
+export const STATUS_CODE_SUCCESS = 200;
+export const STATUS_CODE_CONFLICT = 409;
+
+export const DIFFICULTY_LEVELS = [
+    {
+        value: 1,
+        label: "Easy",
+    },
+    {
+        value: 2,
+        label: "Medium",
+    },
+    {
+        value: 3,
+        label: "Hard",
+    },
+];
